feat: Add login from token on reload. Add logout button in new profile submenu

+++ language_frontend/src/app/services/auth.service.ts
@@ -1,7 +1,8 @@
-import { Injectable } from "@angular/core";
+import { Injectable, PLATFORM_ID, inject } from "@angular/core";
 import { HttpClient } from "@angular/common/http";
 import { Observable, tap } from "rxjs";
 import { AuthRequest, LoginResponse, User } from 'src/app/entities/user.model';
+import { isPlatformBrowser } from "@angular/common";
 
 @Injectable({
   providedIn: 'root'
@@ -11,6 +12,8 @@ export class AuthService {
   private apiUrl = 'http://localhost:4000/api';
   constructor(private http: HttpClient) { };
 
+  private platformId = inject(PLATFORM_ID);
+
   login(params: AuthRequest): Observable<LoginResponse> {
     return this.http.post<LoginResponse>(`${this.apiUrl}/login`, params)
       .pipe(
@@ -19,22 +22,32 @@ export class AuthService {
   };
 
   logout(): void {
-    localStorage.removeItem('access_token');
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.removeItem('access_token');
+    }
   };
 
   getAccessToken(): string | null {
-    return localStorage.getItem('access_token');
+    if (isPlatformBrowser(this.platformId)) {
+      return localStorage.getItem('access_token');
+    }
+
+    return null;
   };
 
   storeToken(response: LoginResponse): void {
-    localStorage.clear();
-    localStorage.setItem('access_token', response.access_token)
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+      localStorage.setItem('access_token', response.access_token)
+    }
   };
 
   signup(params: AuthRequest): Observable<User> {
     return this.http.post<User>(`${this.apiUrl}/signup`, params);
   };
 
-
+  loginFromToken(): Observable<User> {
+    return this.http.post<User>(`${this.apiUrl}/tokenlogin`, null);
+  };
 };
